Rename security_advisories directory to match agent name so alias hash updates on deploy
