+++ agents/security-advisories/__init__.py
@@ -1,6 +1,6 @@
 # Copyright OpenSearch Contributors
 # SPDX-License-Identifier: Apache-2.0
 
-from agents.security_advisories.agent import SecurityAdvisoriesAgent
+from .agent import SecurityAdvisoriesAgent
 
 __all__ = ["SecurityAdvisoriesAgent"]
