+++ agents/security-advisories/lambda/aws_utils.py
