+++ agents/security-advisories/lambda/lambda_function.py
