+++ agents/security-advisories/lambda/projects_handler.py
