+++ agents/security-advisories/lambda/query_utils.py
