+++ agents/security-advisories/lambda/response_builder.py
