+++ agents/security-advisories/lambda/response_filter.py
