+++ agents/security-advisories/lambda/vulnerabilities_handler.py
