+++ tests/agents/security_advisories/lambda/test_dsl_query_builder.py
@@ -18,7 +18,7 @@ from unittest.mock import MagicMock, patch
 
 # Path to the real dsl_query_builder module
 _LAMBDA_PATH = os.path.join(
-    os.path.dirname(__file__), '..', '..', '..', '..', 'agents', 'security_advisories', 'lambda',
+    os.path.dirname(__file__), '..', '..', '..', '..', 'agents', 'security-advisories', 'lambda',
 )
