+++ tests/agents/security_advisories/lambda/test_dsl_query_builder_properties.py
@@ -18,7 +18,7 @@ from hypothesis import strategies as st
 
 # Path to the real lambda module
 _LAMBDA_PATH = os.path.join(
-    os.path.dirname(__file__), '..', '..', '..', '..', 'agents', 'security_advisories', 'lambda',
+    os.path.dirname(__file__), '..', '..', '..', '..', 'agents', 'security-advisories', 'lambda',
 )
