+++ tests/agents/security_advisories/lambda/test_tickets_query_builder.py
@@ -17,7 +17,7 @@ from unittest.mock import MagicMock, patch
 
 # Path to the real tickets_query_builder module
 _LAMBDA_PATH = os.path.join(
-    os.path.dirname(__file__), '..', '..', '..', '..', 'agents', 'security_advisories', 'lambda',
+    os.path.dirname(__file__), '..', '..', '..', '..', 'agents', 'security-advisories', 'lambda',
 )
