+++ tests/agents/security_advisories/test_lambda_function.py
@@ -16,7 +16,7 @@ from unittest.mock import MagicMock, patch
 
 # Path to the real lambda_function module
 _LAMBDA_PATH = os.path.join(
-    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security_advisories', 'lambda',
+    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security-advisories', 'lambda',
 )
