+++ tests/agents/security_advisories/test_map_age_and_neglected_url_call.py
@@ -19,7 +19,7 @@ import pytest
 
 # Path to the real vulnerabilities_handler module
 _LAMBDA_PATH = os.path.join(
-    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security_advisories', 'lambda',
+    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security-advisories', 'lambda',
 )
