+++ tests/agents/security_advisories/test_projects_handler.py
@@ -15,7 +15,7 @@ from unittest.mock import MagicMock, patch
 
 # Path to the real projects_handler module
 _LAMBDA_PATH = os.path.join(
-    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security_advisories', 'lambda',
+    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security-advisories', 'lambda',
 )
