+++ tests/agents/security_advisories/test_resolve_version_tag.py
@@ -14,7 +14,7 @@ import sys
 from unittest.mock import MagicMock, patch
 
 _LAMBDA_PATH = os.path.join(
-    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security_advisories', 'lambda',
+    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security-advisories', 'lambda',
 )
