+++ tests/agents/security_advisories/test_tickets_handler.py
@@ -22,7 +22,7 @@ from hypothesis import strategies as st
 
 # Path to the real lambda module
 _LAMBDA_PATH = os.path.join(
-    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security_advisories', 'lambda',
+    os.path.dirname(__file__), '..', '..', '..', 'agents', 'security-advisories', 'lambda',
 )
